Use sets & canonical paths to avoid O(N) lookups in processed header list
Canonical paths are absolute paths with no dot/dot-dot notations. This
ensures the format is consistent for all forms referencing that file.

Switching over to sets significantly improves the perf of this tool when
used with larger codebase.

Also, reported file path is now the generic version, so slashes are consistent.